Move entire Content Editor form into React app
After this change all the content editor behaviour is governed by the
React app rather than having part of it be an HTML form and part of it
in React.

The downside is JS is now required to edit content; however, that is a
reasonable requirement and the plaintext editor would not have been very
useful anyway.

The new React editor supports IDs that match to the project filenames
used by content creators. There is also auto-generation of URLs from IDs.

Additionally the dependency of htmlminifier has been removed and
replaced with htmlclean, which is a much smaller dependency and seems to
work better for client-side whitespace normalization of HTML
(htmlminifier pulls in a lot of extra dependencies, e.g. UglifyJS, which
do not translate well to client-side use with ES modules).