Update directory to use new layout technique
Instead of using many floats, :first-child, and adjacency selectors, the
directory is now laid out using display: table and table-cell. This
makes it a bit more flexible for auto-scaling and works better in IE8,
however it may work worse in IE7. Since the site didn't work at all in
IE7 previously that doesn't seem like a loss.

The appearance of the directory has also been updated to be a bit more
consistent.